requirements.txt
python-dotenv==1.0.1
requests==2.32.3
web3==6.19.0
websockets==12.0
websocket-client==1.8.0
numpy==1.26.4
eth-abi==4.2.1
fastapi==0.111.0
uvicorn==0.30.0
